update E2E test
