+++ .github/workflows/ci-cd.yml
@@ -47,7 +47,7 @@ jobs:
         run: BUILD_BRANCH=$(echo "${GITHUB_REF#refs/heads/}") BUILD_COMMIT=${{ github.sha }} npm run build
 
       - name: Store assets
-        if: ${{ github.event_name == 'push' && (github.ref == 'refs/heads/development' || github.ref == 'refs/heads/main') }}
+        if: ${{ github.event_name == 'push' && (github.ref == 'refs/heads/consumer_mode_timeout' || github.ref == 'refs/heads/main') }}
         uses: actions/upload-artifact@v3
         with:
           name: assets
@@ -58,7 +58,7 @@ jobs:
     name: Upload assets
     runs-on: ubuntu-20.04
     needs: build
-    if: ${{ github.event_name == 'push' && github.ref == 'refs/heads/development' }}
+    if: ${{ github.event_name == 'push' && github.ref == 'refs/heads/consumer_mode_timeout' }}
     strategy:
       matrix:
         environment:
