+++ package.json
@@ -64,7 +64,7 @@
   "bugs": "https://github.com/splitio/javascript-browser-client/issues",
   "homepage": "https://github.com/splitio/javascript-browser-client#readme",
   "dependencies": {
-    "@splitsoftware/splitio-commons": "1.8.3-rc.2",
+    "@splitsoftware/splitio-commons": "1.8.3-rc.3",
     "@types/google.analytics": "0.0.40",
     "unfetch": "^4.2.0"
   },
